412 - Make timestamp generator monotonic
Split platform specific get time implementations into platform specific
files.
